project_root/requirements.txt
numpy==2.1.3
scipy==1.14.1
pandas==2.2.3
matplotlib==3.9.2
scikit-learn==1.5.2
wfdb==4.3.1
neurokit2==0.2.10
pyhrv==0.4.1
heartpy==1.2.7
tqdm==4.66.5
pyyaml==6.0.2
seaborn==0.13.2
jupyter==1.1.1
joblib==1.4.2
statsmodels==0.14.4
xgboost==2.1.2
lightgbm==4.5.0
ucimlrepo==0.0.7
requests==2.32.3
beautifulsoup4==4.12.3
